apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Nexus Console API - FastAPI Backend"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.0",
    "asyncpg>=0.31.0",
    "celery>=5.6.0",
    "fastapi>=0.127.0",
    "httpx>=0.28.0",
    "pydantic[email]>=2.12.0",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "sqlalchemy[asyncio]>=2.0.45",
    "uvicorn[standard]>=0.40.0",
    "structlog>=25.5.0",
    "slowapi>=0.1.9",
    "fastapi-cache2>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.15.0",
]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[dependency-groups]
dev = [
    "aiosqlite>=0.22.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
